requirements.txt
2captcha-python==1.1.3
alembic==1.9.4
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
exceptiongroup==1.1.0
filelock==3.9.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dotenv==0.21.1
python-telegram-bot==20.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rfc3986==1.5.0
schedule==1.1.0
Scrapy==2.8.0
selenium==4.8.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.4
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.14
w3lib==2.1.1
wsproto==1.2.0
zope.interface==5.5.2
